Extract ignore rule loading partial
